A program created using processing that simulates a traffic light intersection
A car will travel alongside a road, and will respond to traffic lights.
When the light is green, it will go, yellow, slow down, and stop at red.
The user can interact with the program by engaging a pedestrian
crosswalk. The program will check if any cars are in the way, and if it
is safe to cross, then the pedestrain (represented by a blue dot) will
begin to cross the street.